% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignBins.R
\name{alignBins}
\alias{alignBins}
\title{Removes sequences shorter than a given cutoff}
\usage{
alignBins(all_results, config)
}
\arguments{
\item{all_results}{A list of all results given the class 'all_results'.}

\item{config}{A list of configuration options}
}
\description{
Removes sequences shorter than a given cutoff
}

