% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignBins.R
\name{alignBins_internal}
\alias{alignBins_internal}
\title{aligns fwd and rev sequences to a profile}
\usage{
alignBins_internal(cur_fwd_seqs, cur_rev_seqs, profile_file, working_dir,
  pid)
}
\description{
a custom guide tree is constructed to ensure that the alignment of the fwd
reads does not interfere with the alignment of the rev reads and vice versa.
}
