% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignBinsMSA.R
\name{alignBinsMSA_internal}
\alias{alignBinsMSA_internal}
\title{aligns fwd and rev sequences to a profile}
\usage{
alignBinsMSA_internal(cur_seqs, working_dir, pid)
}
\description{
a custom guide tree is constructed to ensure that the alignment of the fwd
reads does not interfere with the alignment of the rev reads and vice versa.
}

