% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignBinsSP.R
\name{alignBinsSP_internal}
\alias{alignBinsSP_internal}
\title{aligns fwd and rev sequences to a profile}
\usage{
alignBinsSP_internal(cur_seqs, profile_seqs, working_dir, pid, which_pair)
}
\description{
a custom guide tree is constructed to ensure that the alignment of the fwd
reads does not interfere with the alignment of the rev reads and vice versa.
}

