% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allResults.R
\name{allResults}
\alias{allResults}
\title{Performs any final processing operations}
\usage{
allResults(all_results, config)
}
\arguments{
\item{all_results}{A list of all results given the class 'all_results'.}

\item{config}{A list of configuration options}
}
\description{
Performs any final processing operations
}

