% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ambigSeqs.R
\name{ambigSeqs}
\alias{ambigSeqs}
\title{Removes sequences with too many ambig bases}
\usage{
ambigSeqs(all_results, config)
}
\arguments{
\item{all_results}{A list of all results given the class 'all_results'.}

\item{config}{A list of configuration options}
}
\description{
Removes sequences with too many ambig bases
}
