% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processPrimers.R
\name{applyOperation}
\alias{applyOperation}
\title{Applies a single operation}
\usage{
applyOperation(operation, all_results, config)
}
\arguments{
\item{operation}{The name of the operation to apply. A function by this name
with the arguments all_results and config must exists. Additionally, the
functions saveToDisk, genReport, genSummary and print must be specialized
for objects of the class matching the name of the operation.}

\item{all_results}{A list of all results given the class 'all_results'.}

\item{config}{A list of configuration options}
}
\description{
Applies a single operation
}

