% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binSeqErr.R
\name{binSeqErr}
\alias{binSeqErr}
\title{Computes the sequencing parameters for each bin}
\usage{
binSeqErr(all_results, config)
}
\arguments{
\item{all_results}{A list of all results given the class 'all_results'.}

\item{config}{A list of configuration options}
}
\description{
Computes the sequencing parameters for each bin
}

