% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binSeqErr.R
\name{binSeqErr_internal}
\alias{binSeqErr_internal}
\title{Computes the sequencing error rate for each bin}
\usage{
binSeqErr_internal(cons_dat, msa_dat)
}
\arguments{
\item{cons_dat}{The consensus sequences. Their names must start with the PID
followed by an underscore and then anything.}

\item{msa_dat}{The reads with sequences of each bin aligned. The sequence
name must contain PID:_ followed by the PID somewhere in the name.}
}
\description{
Under the assumption that there was zero PCR recombination and no PID
collisions, the only source of heterogeneity in the bins are sequencing
errors. Thus a comparison between each read in a bin and the bin's consensus
sequence can be taken as an indication of the sequencing error rates.
}
\details{
A full characterization of the errors is returned that tallies each match
and mismatch and the quality at which they happened.
}
