% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensus_cutoff.R
\name{build_cutoff_models}
\alias{build_cutoff_models}
\title{Builds the consensus cutoff models given a dataset index file}
\usage{
build_cutoff_models(index_file, verbose = FALSE, models_file = NULL)
}
\arguments{
\item{index_file}{The file that describes the available datasets}

\item{verbose}{Prints progress and fitting informations and plots models to
active graphics device.}

\item{models_file}{If not NULL, then the models data.frame will be written
to this file.}
}
\description{
Builds the consensus cutoff models given a dataset index file
}
