% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeMetrics.R
\name{computeMetrics}
\alias{computeMetrics}
\title{Computes metrics useful for reporting}
\usage{
computeMetrics(result, config, seq_dat)
}
\arguments{
\item{result}{The result produced from a specific operation}

\item{config}{A list of configuration options}
}
\description{
Computes metrics useful for reporting
}
