% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doubleClusterTally.R
\name{doubleClusterTally}
\alias{doubleClusterTally}
\title{Tallies to true sizes of clusters of clusters}
\usage{
doubleClusterTally(all_results, config)
}
\arguments{
\item{all_results}{A list of all results given the class 'all_results'.}

\item{config}{A list of configuration options}
}
\description{
Tallies to true sizes of clusters of clusters
}
