% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractPIDs.R
\name{extractPIDs}
\alias{extractPIDs}
\title{Extracts the PIDs after affixes were trimmed and add then to the sequence names}
\usage{
extractPIDs(all_results, config)
}
\arguments{
\item{all_results}{A list of all results given the class 'all_results'.}

\item{config}{A list of configuration options}
}
\description{
Extracts the PIDs after affixes were trimmed and add then to the sequence names
}

