% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ambigSeqs.R
\name{genKeptVector}
\alias{genKeptVector}
\title{Generates a vector of T/F indicating which sequences were kept or trimmed.}
\usage{
genKeptVector(result)
}
\description{
Generates a vector of T/F indicating which sequences were kept or trimmed.
}

