% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genReport.R
\name{genReport}
\alias{genReport}
\title{Generates a report for a operation or the full process}
\usage{
genReport(result, config)
}
\arguments{
\item{result}{The result produced from a specific operation}
}
\description{
Generates a report for a operation or the full process
}

