% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genSummary.R
\name{genSummary}
\alias{genSummary}
\title{Generates a summary for a operation or the full process}
\usage{
genSummary(result, config, seq_dat)
}
\arguments{
\item{result}{The result produced from a specific operation}
}
\description{
Generates a summary for a operation or the full process
}

