% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genSummary.R
\name{genSummary_case3}
\alias{genSummary_case3}
\title{Generates summary table when there is one trim step with many values}
\usage{
genSummary_case3(result, config, seq_dat)
}
\arguments{
\item{result}{The result produced from a specific operation}

\item{config}{A list of configuration options}
}
\description{
Generates summary table when there is one trim step with many values
}
