% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genSummary.R
\name{genSummary_case4}
\alias{genSummary_case4}
\title{Generates summary table when there is many trim steps each with many values}
\usage{
genSummary_case4(result, config, seq_dat, round_digits = 0)
}
\arguments{
\item{result}{The result produced from a specific operation}
}
\description{
Generates summary table when there is many trim steps each with many values
}

