% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genSummary.R
\name{genSummary_internal}
\alias{genSummary_internal}
\title{Computes the summary tables of an operation}
\usage{
genSummary_internal(operation, parameters, kept_seq_dat, trimmed_seq_dat)
}
\description{
Computes the summary tables of an operation
}

