% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensus_cutoff.R
\name{get_consensus_cutoff}
\alias{get_consensus_cutoff}
\title{Returns the consensus cutoff to use based on a given models file}
\usage{
get_consensus_cutoff(pl, main_offspring_size, models, phi = 1/50)
}
\arguments{
\item{pl}{PID length (9)}

\item{main_offspring_size}{The size of the largest bin in the dataset
produced by the binner. NB! This must be the size of the bin BEFORE outlier
removal}

\item{models}{A data.frame containing the parameters for various models. As
produced by build_cutoff_models.}

\item{phi}{Sequencing read error rate (1/50)}
}
\description{
This function will find the model that matches your input parameters and use
it to predict the appropriate consensus cutoff for your dataset
}

