% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensus_cutoff.R
\name{get_cutoff_models}
\alias{get_cutoff_models}
\title{Returns a data.frame of the stored models}
\usage{
get_cutoff_models()
}
\description{
A number of consensus cutoff models have been pre-built. This function
returns a data.frame with there parameters and the parameters that describes
the datasets that they are applicable to
}
