% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binSeqErr.R
\name{group_cons_bins_internal}
\alias{group_cons_bins_internal}
\title{Group consensus sequences and their read together}
\usage{
group_cons_bins_internal(cons_dat, msa_dat)
}
\arguments{
\item{cons_dat}{The consensus sequences. Their names must start with the PID
followed by an underscore and then anything.}

\item{msa_dat}{The reads with sequences of each bin aligned. The sequence
name must contain PID:_ followed by the PID somewhere in the name.}
}
\description{
Constructs a list with a single element for each bin. This element is a list
that contains the consensus sequence as well as the aligned reads of the
bins.
}

