% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genSummary.R
\name{kable_summary}
\alias{kable_summary}
\title{formats a summary table for markdown}
\usage{
kable_summary(summary_tab)
}
\arguments{
\item{summary_tab}{A data.frame as produced by genSummary}
}
\description{
formats a summary table for markdown
}

