% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeIUPACScoreMat.R
\name{makeIUPACScoreMat}
\alias{makeIUPACScoreMat}
\title{This function creates a scoring matrix for aligning IUPAC sequences}
\usage{
makeIUPACScoreMat()
}
\description{
This function creates a scoring matrix for aligning IUPAC sequences
}
