% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensus_cutoff.R
\name{moving_statistic}
\alias{moving_statistic}
\title{Computes a moving statistic over a sliding window.}
\usage{
moving_statistic(y, x, statistic, bw = 5, n_steps = 30, ...)
}
\description{
Computes a moving statistic over a sliding window.
}
