% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensus_cutoff.R
\name{mutate_letter}
\alias{mutate_letter}
\title{Mutates a nucleotide to another}
\usage{
mutate_letter(x)
}
\arguments{
\item{x}{The letter to mutate}
}
\description{
Mutates a nucleotide to another
}
