% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwiseMSA.R
\name{pairwiseMSA}
\alias{pairwiseMSA}
\title{Generates a MSA relative to a profile without any inserts into the profile}
\usage{
pairwiseMSA(reads, ref_seq)
}
\arguments{
\item{reads}{Reads either as a ShortRead or DNAStringSet object with the @sread and @id fields.}

\item{ref_seq}{The profile sequence to use as a reference}
}
\description{
Sequence quality information will be ignored.
}
