% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensus_cutoff.R
\name{plot_consensus_cutoff_model}
\alias{plot_consensus_cutoff_model}
\title{Plots a consensus cutoff model}
\usage{
plot_consensus_cutoff_model(a, b, d, e)
}
\arguments{
\item{a}{The a parameter for the model}

\item{b}{The b parameter for the model}

\item{d}{The d parameter for the model}

\item{e}{The e parameter for the model}
}
\description{
The following model is assumed: y = a + bx + dx^e.
}
