% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensus_cutoff.R
\name{plot_consensus_cutoff_sim}
\alias{plot_consensus_cutoff_sim}
\title{Plots the results of a consensus_cutoff simulation}
\usage{
plot_consensus_cutoff_sim(dat)
}
\arguments{
\item{dat}{The data simulated by sim_all_parents_main_off}
}
\description{
Plots the results of a consensus_cutoff simulation
}
