% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepConfig.R
\name{prepConfig}
\alias{prepConfig}
\title{Prepares config list for inclusion in final report}
\usage{
prepConfig(all_results, config)
}
\arguments{
\item{all_results}{A list of all results given the class 'all_results'.}

\item{config}{A list of configuration options}
}
\description{
Converts the configuration used to call MotifBinner to a data.frame that can
be included in the final report.
}
\details{
The configuration used to specify which opertation should be applied and
what settings to use is contained in a nested list of lists. These lists are
typically built using the buildConfig function. By including these settings
and the version of MotifBinner in the final report an easy way to document
the exact procedures that were used to produce a dataset.

The main action performed by this prepConfig operation is just to convert
the nested list of lists that holds the config to a data.frame and to
provide a convenient way, using the standard approach of MotifBinner, to add
this information into the final report.
}
