% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processPrimers.R
\name{processPrimers}
\alias{processPrimers}
\title{Processes a dataset with Primer IDs}
\usage{
processPrimers(fwd_reads_file = NULL, rev_reads_file = NULL,
  output_dir = NULL, prefix_for_names = NULL,
  operation_list = c("loadData"), intermediate_reports = TRUE,
  verbosity = 0, report_type = c("html", "pdf"), ...)
}
\arguments{
\item{fwd_reads_file}{The name of the fastq file with the forward reads.}

\item{rev_reads_file}{The name of the fastq file with the reverse reads.}

\item{output_dir}{The directory in which the output must be produced.}

\item{prefix_for_names}{The basename to use for naming output.}

\item{operation_list}{The list of operations to apply to the input data.}

\item{intermediate_reports}{Should intermediate reports be produced after
each operations? (Useful for debugging)}

\item{verbosity}{The output level. 0 = no output. 1 = info about which
operation is currently running. 2 = info about progress of current
operation. 3 = super verbose debugging output.}

\item{report_type}{vector of types of reports to procude. Valid options:
'html', 'pdf'.}
}
\description{
Applies a series of operations to the input dataset that generates consensus
sequences from raw (or preprocessed) datasets containing Primer IDs.
}
