% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualTrim.R
\name{qualTrim}
\alias{qualTrim}
\title{Removes sequences that does not match specified quality}
\usage{
qualTrim(all_results, config)
}
\arguments{
\item{all_results}{A list of all results given the class 'all_results'.}

\item{config}{A list of configuration options}
}
\description{
Removes sequences that does not match specified quality
}
