% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binSeqErr.R
\name{reformat_tallies}
\alias{reformat_tallies}
\title{Reformats the tallies produced by tallyPrimerSeqError_cpp}
\usage{
reformat_tallies(tallies_list)
}
\arguments{
\item{tallies_list}{The tallies list produced by tallyPrimerSeqErrors_cpp}
}
\value{
A data.frame with the tallies of the sequencing parameters
}
\description{
Reformats the tallies produced by tallyPrimerSeqError_cpp
}
