% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeChars.R
\name{removeChars}
\alias{removeChars}
\title{Removes all instances of a specified character from a set of sequences}
\usage{
removeChars(all_results, config)
}
\arguments{
\item{all_results}{A list of all results given the class 'all_results'.}

\item{config}{A list of configuration options}
}
\description{
Note that the current consensus building operation assigns a quality of zero
to both gaps and N's. These are the two primary types of characters to
remove and hence it does not make sense to report on the quality of the
letters that were removed.
}
