% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveToDisk.R
\name{saveToDisk}
\alias{saveToDisk}
\title{Save the relevant results of an operation to disk}
\usage{
saveToDisk(result, config)
}
\arguments{
\item{result}{The result produced from a specific operation}

\item{config}{A list of configuration options}
}
\description{
Save the relevant results of an operation to disk
}

