% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqLength.R
\name{seqLength}
\alias{seqLength}
\title{Removes perimer-dimer sequences (based on sequence length)}
\usage{
seqLength(all_results, config)
}
\arguments{
\item{all_results}{A list of all results given the class 'all_results'.}

\item{config}{A list of configuration options}
}
\description{
Removes perimer-dimer sequences (based on sequence length)
}

