% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processBadPIDs.R
\name{shortReadQ_forced_append}
\alias{shortReadQ_forced_append}
\title{Forces ShortReadQ objects to append into a single ShortReadQ object}
\usage{
shortReadQ_forced_append(x)
}
\arguments{
\item{x}{list of ShortReadQ objects}
}
\description{
Forces ShortReadQ objects to append into a single ShortReadQ object
}
