% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensus_cutoff.R
\name{sim_all_parents_main_off}
\alias{sim_all_parents_main_off}
\title{Simulate range of parents and their main offspring sizes}
\usage{
sim_all_parents_main_off(ns = seq(100, 10000, 100), pl = 9,
  phi = (1/50), repeats = 1000, verbose = FALSE, ncpu = 8)
}
\arguments{
\item{pl}{PID length (9)}

\item{phi}{Sequencing read error rate (1/50)}

\item{repeats}{The number of repeats to run for each bin size}

\item{ncpu}{Number of cores to use for simulation}

\item{n}{a vector of bin_sizes (seq(100, 10000, 100))}

\item{vebose}{If TRUE print loop progress}
}
\description{
Simulate range of parents and their main offspring sizes
}
