% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensus_cutoff.R
\name{sim_one_parent_main_off}
\alias{sim_one_parent_main_off}
\title{Simulate a parent and its main offspring size}
\usage{
sim_one_parent_main_off(n = 1000, pl = 9, phi = (1/50))
}
\arguments{
\item{n}{bin size (1000)}

\item{pl}{PID length (9)}

\item{phi}{Sequencing read error rate (1/50)}
}
\description{
Simulate a parent and its main offspring size
}

