% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensus_cutoff.R
\name{sim_wrapper}
\alias{sim_wrapper}
\title{Wrapper that performs simulation and process the results}
\usage{
sim_wrapper(output_dir, file_name, size_start = 100, size_stop = 10000,
  size_step = 100, pl = 9, phi = (1/50), repeats = 1000,
  verbose = FALSE, ncpu = 8)
}
\arguments{
\item{output_dir}{The output directory where the simulation result must be
placed}

\item{file_name}{The name of the output file in which the results will be
stored. Must end in .csv.}

\item{size_start}{The smallest bin size to simulate}

\item{size_stop}{The largest bin size to simulate}

\item{size_step}{The size of the steps to take to reach the largest bin size}

\item{pl}{PID length (9)}

\item{phi}{Sequencing read error rate (1/50)}

\item{repeats}{The number of repeats to run for each bin size}

\item{ncpu}{Number of cores to use for simulation}

\item{vebose}{If TRUE print loop progress}
}
\description{
The function calls the sim_all_parents_main_off function and saves the
result to a folder together with a file that contains a description of the
simulation parameters.
}
