% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timing.R
\name{sliceTime}
\alias{sliceTime}
\title{Internal function that reforms a list storing exe times}
\usage{
sliceTime(x)
}
\arguments{
\item{x}{a list of execution times}
}
\description{
Internal function that reforms a list storing exe times
}
