% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timing.R
\name{timingTable}
\alias{timingTable}
\title{Produces the timing tables in the reports}
\usage{
timingTable(result)
}
\arguments{
\item{result}{A result object with a timing element}
}
\description{
Produces the timing tables in the reports
}
