% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trimAffixes.R
\name{trimAffixes}
\alias{trimAffixes}
\title{Trims the ends of the sequences so that they all start at the same place}
\usage{
trimAffixes(all_results, config)
}
\arguments{
\item{all_results}{A list of all results given the class 'all_results'.}

\item{config}{A list of configuration options}
}
\description{
Trims the ends of the sequences so that they all start at the same place
}
