% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractData.R
\name{extractData}
\alias{extractData}
\title{Extracts required dataset by navigating down into a list}
\usage{
extractData(all_results, config)
}
\arguments{
\item{all_results}{A list of all results given the class 'all_results'.}

\item{config}{A list of configuration options}
}
\description{
Specify the names of the elements you want to extract using 
the extract_levels parameter. If you have results stashed
in a list like this list('X' = list('bla' = the_target_data)),
then specify the extract levels as extract_levels = c('X', 'bla')
}
