% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadData.R
\name{loadData}
\alias{loadData}
\title{Loads the fwd and rev fastq reads}
\usage{
loadData(all_results, config)
}
\arguments{
\item{all_results}{A list of all results given the class 'all_results'.}

\item{config}{A list of configuration options}
}
\description{
Loads the fwd and rev fastq reads
}

