% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processBadPIDs.R
\name{shortReadQ_forced_append}
\alias{shortReadQ_forced_append}
\title{Forces two ShortReadQ objects to append}
\usage{
shortReadQ_forced_append(x, y)
}
\arguments{
\item{x}{ShortReadQ object}

\item{y}{ShortReadQ object}
}
\description{
Forces two ShortReadQ objects to append
}

